+++ app/controllers/sync_controller.rb
@@ -1,13 +1,7 @@
 class SyncController < ApplicationController
   
-  # Require Google calendar api.
-  require 'google/apis/calendar_v3'
-  
-  # Require the syncing code.
-  require 'syncing/sync'
-
   # Include syncing code from module.
-  include SYNC
+  include Syncing::GoogleCalendar
   
   # Sync data from Google calendars
   def sync
